Add modify filters commands

+++ src/Commands/Input/SubscriptionInput.cs
@@ -38,7 +38,7 @@ namespace WhMgr.Commands.Input
             var pokemonMessage = (await _context.RespondEmbed("Enter either the Pokemon name(s) or Pokedex ID(s) separated by a comma to subscribe to (i.e. Mewtwo,Dragonite):", DiscordColor.Blurple)).FirstOrDefault();
             var pokemonSubs = await _context.WaitForUserChoice();
             // Validate the provided pokemon list
-            var validation = PokemonValidation.Validate(pokemonSubs, (int)maxPokemonId);
+            var validation = PokemonValidation.Validate(pokemonSubs, maxPokemonId);
             if (validation == null || validation.Valid.Count == 0)
             {
                 await _context.RespondEmbed(Translator.Instance.Translate("NOTIFY_INVALID_POKEMON_IDS_OR_NAMES").FormatText(_context.User.Username, string.Join(", ", validation.Invalid)), DiscordColor.Red);
